Role validation added, cors enabled

+++ server/src/controllers/reservation.js
@@ -1,25 +1,38 @@
 "use strict";
 const Reservation = require("../models/reservation");
+const helpers = require('../helpers/validateHelper');
+const accessControl = require('../helpers/accessControl');
 
 const controller = {
   list: async (request, response) => {
-    try {
-      const query = typeof request.params.storeId !== "undefined"
-          ? { storeId: request.params.storeId }
-          : {};
-      const reservations = await Reservation.find(query)
-        .populate({
-          path: "storeId",
-          model: "Store",
-        })
-        .populate({
-          path: "serviceId",
-          model: "Service",
+    const allowedRoles = await helpers.getMongoRoles(accessControl.permissions.reservations.get);
+    if(helpers.compareRoles(allowedRoles, request.userRoles)) {
+      try {
+        const query = typeof request.params.storeId !== "undefined"
+            ? { storeId: request.params.storeId }
+            : {};
+        const reservations = await Reservation.find(query)
+          .populate({
+            path: "clientId",
+            model: "User",
+          })
+          .populate({
+            path: "storeId",
+            model: "Store",
+          })
+          .populate({
+            path: "serviceId",
+            model: "Service",
+          });
+        return response.status(200).send(reservations);
+      } catch (error) {
+        return response.status(400).send({
+          message: error.message
         });
-      return response.status(200).send(reservations);
-    } catch (error) {
-      return response.status(400).send({
-        message: error.message
+      }
+    } else {
+      return response.status(403).send({
+        message: "Forbidden"
       });
     }
   },
@@ -51,35 +64,49 @@ const controller = {
   },
 
   update: async (request, response) => {
-    try {
-      const updateReservation = request.params.id;
-      const params = request.body;
-      const options = {new: true};
-  
-      await Reservation.findByIdAndUpdate(updateReservation, params, options)
-        return response.status(200).send({
-          message: "Success Reservation Updated",
-          reservation: params,
+    const allowedRoles = await helpers.getMongoRoles(accessControl.permissions.reservations.put);
+    if(helpers.compareRoles(allowedRoles, request.userRoles)) {
+      try {
+        const updateReservation = request.params.id;
+        const params = request.body;
+        const options = {new: true};
+    
+        await Reservation.findByIdAndUpdate(updateReservation, params, options)
+          return response.status(200).send({
+            message: "Success Reservation Updated",
+            reservation: params,
+          });
+      } catch (error) {
+        return response.status(400).send({
+          status: "Id Error",
+          message: error.message
         });
-    } catch (error) {
-      return response.status(400).send({
-        status: "Id Error",
-        message: error.message
+      }
+    } else {
+      return response.status(403).send({
+        message: "Forbidden"
       });
     }
   },
 
   delete: async (request, response) => {
-    try {
-      const deleteReservation = await Reservation.findByIdAndDelete(request.params.id);
-      return response.status(200).send({
-        message: "Success Reservation Deleted",
-        reservation: deleteReservation,
-      });
-    } catch (error) {
-      return response.status(400).send({
-        status: "Id Error",
-        message: error.message
+    const allowedRoles = await helpers.getMongoRoles(accessControl.permissions.reservations.delete);
+    if(helpers.compareRoles(allowedRoles, request.userRoles)) {
+      try {
+        const deleteReservation = await Reservation.findByIdAndDelete(request.params.id);
+        return response.status(200).send({
+          message: "Success Reservation Deleted",
+          reservation: deleteReservation,
+        });
+      } catch (error) {
+        return response.status(400).send({
+          status: "Id Error",
+          message: error.message
+        });
+      }
+    } else {
+      return response.status(403).send({
+        message: "Forbidden"
       });
     }
   }
